requirements.txt
acres==0.1.0
asttokens==3.0.0
bitarray==3.0.0
camelot-py==1.0.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.1.0
fonttools==4.55.3
idna==3.10
ipython==8.31.0
jedi==0.19.2
jpype1==1.5.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
msal==1.31.1
msoffcrypto-tool==5.4.2
numpy==2.2.1
Office365-REST-Python-Client==2.5.14
olefile==0.47
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20240706
pdfreader==0.1.15
pillow==11.1.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.1
pyparsing==3.2.1
pypdf==5.2.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
tabula-py==2.10.0
tabulate==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
unicode==2.9
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
xlrd==2.0.1
XlsxWriter==3.2.0
